app/requirements.txt
cmake==3.30.4
fastapi==0.115.2
#fastui==0.7.0
#huggingface-hub==0.25.2
#langchain==0.3.3
#langchain-community==0.3.2
#langchain-core==0.3.10
#langchain-text-splitters==0.3.0
llama_cpp_python==0.3.1
#pandas==2.2.3
pydantic-settings==2.5.2
qdrant-client==1.12.0
scikit-learn==1.5.2
setuptools==75.2.0
sse-starlette==2.1.3
uvicorn==0.32.0
python-dotenv
litellm